# libs/api-client/ng-api-connector/src/lib/generated/model/changePasswordResultDto.ts
/* eslint-disable no-control-regex */
/* eslint-disable @typescript-eslint/no-namespace */
/* eslint-disable @typescript-eslint/no-unused-vars */
/* eslint-disable @typescript-eslint/no-explicit-any */
/* eslint-disable @typescript-eslint/no-empty-interface */
/**
 * API Documentation
 * API Description.<br/><br/>Download the api docs file in JSON format <a target=\"_blank\" href=\"/api/docs/swagger.json\">here</a>.<br/><br/>Download the api docs file in YAML format <a target=\"_blank\" href=\"/api/docs/swagger.yaml\">here</a>.
 *
 * OpenAPI spec version: 0.0.2
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ChangePasswordResultDto {
  passwordChanged: boolean;
}